refactor: Restructure crates, bump deps, and add Guile
Major refactor.

This project has been reborn!

With my work on `automon` and `displayz`, AND `guile-rs`, we can
hopefully really make a start on this window manager.

I haven't got a Mac yet, and I use NixOS primarily, so Windows/macOS
might be on the back-burner for a while, but a common interface for the
WM will help when this arrives.

Guile will be used to configure the WM - with active introspection, so
you can actually edit the WM *live*. A la StumpWM.

More to come soon...